requirements.txt
alembic==1.4.0
astroid==2.3.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.6.0
Flask==1.1.1
Flask-CLI==0.4.0
Flask-Cors==3.0.8
flask-marshmallow==0.11.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
idna==2.9
importlib-metadata==1.5.0
inflection==0.3.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Mako==1.1.1
MarkupSafe==1.1.1
marshmallow==3.5.0
marshmallow-sqlalchemy==0.22.2
mccabe==0.6.1
openapi-spec-validator==0.2.8
pylint==2.4.4
pylint-flask==0.6
pylint-plugin-utils==0.6
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
PyYAML==5.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.13
swagger-ui-bundle==0.0.6
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
zipp==3.0.0
